+++ packages/apidom-ls/src/config/openapi/media-type/lint/index.ts
@@ -2,7 +2,14 @@ import allowedFieldsLint from './allowed-fields';
 import schemaTypeLint from './schema--type';
 import examplesValuesTypeLint from './examples--values-type';
 import encodingValuesTypeLint from './encoding--values-type';
+import examplesMutuallyExclusiveLint from './examples--mutually-exclusive';
 
-const lints = [schemaTypeLint, examplesValuesTypeLint, encodingValuesTypeLint, allowedFieldsLint];
+const lints = [
+  schemaTypeLint,
+  examplesValuesTypeLint,
+  encodingValuesTypeLint,
+  allowedFieldsLint,
+  examplesMutuallyExclusiveLint,
+];
 
 export default lints;