+++ packages/apidom-reference/src/resolve/util.ts
@@ -13,7 +13,12 @@ import { ResolverError, UnmatchedResolverError } from '../util/errors';
  */
 // eslint-disable-next-line import/prefer-default-export
 export const readFile = async (file: IFile, options: IReferenceOptions): Promise<Buffer> => {
-  const resolvers: IResolver[] = await plugins.filter('canRead', file, options.resolve.resolvers);
+  const optsBoundResolvers: IResolver[] = options.resolve.resolvers.map((resolver) => {
+    const clonedResolver = Object.create(resolver);
+    return Object.assign(clonedResolver, options.resolve.resolverOpts);
+  });
+
+  const resolvers: IResolver[] = await plugins.filter('canRead', file, optsBoundResolvers);
 
   // we couldn't find any resolver for this File
   if (isEmpty(resolvers)) {
@@ -21,11 +26,7 @@ export const readFile = async (file: IFile, options: IReferenceOptions): Promise
   }
 
   try {
-    const optsBoundResolvers = resolvers.map((resolver) => {
-      const clonedResolver = Object.create(resolver);
-      return Object.assign(clonedResolver, options.resolve.resolverOpts);
-    });
-    const { result } = await plugins.run('read', [file], optsBoundResolvers);
+    const { result } = await plugins.run('read', [file], resolvers);
     return result;
   } catch (error: any) {
     throw new ResolverError(`Error while reading file "${file.uri}"`, error);